YAML sim scene file, skybox, window icon, above water works

+++ camera_faker/include/textShader.hpp
@@ -114,7 +114,7 @@ private:
         glPixelStorei(GL_UNPACK_ALIGNMENT, 1); // disable byte-alignment restriction
         FT_Set_Pixel_Sizes(face, 0, TEXT_HEIGHT);
 
-                for (unsigned char c = 0; c < 200; c++)
+        for (unsigned char c = 0; c < 200; c++)
         {
             // load character glyph
             if (FT_Load_Char(face, c, FT_LOAD_RENDER))
@@ -146,7 +146,7 @@ private:
                 texture,
                 glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
                 glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-                face->glyph->advance.x};
+                static_cast<unsigned int>(face->glyph->advance.x)};
             Characters.insert(std::pair<char, Character>(c, character));
         }
         // Free resources
